finalized renderpass setup

+++ Engine/Scene/Camera.cpp
@@ -35,7 +35,7 @@ glm::mat4 Camera::GetViewMatrix() const
 
 glm::mat4 Camera::GetProjectionMatrix() const
 {
-	return glm::perspective(glm::radians(Zoom), Viewport.Width / Viewport.Height, 0.1f, 200.0f);
+	return glm::perspective(glm::radians(Zoom), (float)Viewport.Width / (float)Viewport.Height, 0.1f, 200.0f);
 }
 
 // processes input received from any keyboard-like input system. Accepts input parameter in the form of camera defined ENUM (to abstract it from windowing systems)
